api fix

+++ Bank Application - Assignment/bank-app/bank-app/src/main/java/com/example/bank_app/Account/AccountService.java
@@ -1,4 +1,7 @@
 package com.example.bank_app.Account;
+import com.example.bank_app.User.User;
+import com.example.bank_app.User.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -7,11 +10,13 @@ import java.util.List;
 @Service
 public class AccountService {
 
+    private final UserRepository userRepository;
     private final AccountRepository accountRepository;
 
     @Autowired
-    public AccountService(AccountRepository accountRepository) {
+    public AccountService(AccountRepository accountRepository, UserRepository userRepository) {
         this.accountRepository = accountRepository;
+        this.userRepository = userRepository;
     }
 
     public List<Account> getAllAccounts() {
@@ -26,7 +31,17 @@ public class AccountService {
         return accountRepository.save(account);
     }
 
+    @Transactional
     public void deleteAccount(Long id) {
-        accountRepository.deleteById(id);
+        Account account = accountRepository.findById(id).orElse(null);
+        if (account != null) {
+            User user = account.getUser();
+            if (user != null) {
+                account.setUser(null);
+                accountRepository.save(account);
+                userRepository.deleteById(user.getId());
+            }
+            accountRepository.deleteById(id);
+        }
     }
 }
